Data.Allen.Relation

Description

This module provides functions for working with relations. Note that almost all exposed functions only work with relation bitsets. This is done mainly to optimize the speed in calculations involving relations.

The RelationBits type is a synonym for a 16 bit unsigned integer. Note that since Allen's interval algebra only defines 13 relations, the remaining 3 bits are unused. So the bit representation of every possible relation looks like this:

0b0001111111111111

Modifying the extra 3 bits will not affect the result of any calculations. To view in exact detail how a Relation converted to a bit representation, see the toBits function.

Documentation

converse :: RelationBits -> RelationBits

Return the converse of a Relation bitset.

hasRelation :: Relation -> IntervalID -> IntervalID -> Allen Bool

Return if a relation exists between two intervals.

composeSingle :: Relation -> Relation -> RelationBits

Compose two relations.

compose :: RelationBits -> RelationBits -> RelationBits

Compose two sets of relations.

bitsFromString :: String -> RelationBits

Given a string, return the bit representation of the set of relations. Valid characters: pmoFDseSdfoMP.

You may also use full to represent all relations, or concur to represent all relations excluding Precedes and PrecededBy.