Data.Allen

Description

Module : Data.Allen

Description : Main module for Allen's interval algebra.

Maintainer : Archaversine |

Documentation

module Data.Allen.Types

module Data.Allen.Interval

module Data.Allen.Relation

execAllen :: Allen a -> IntervalGraph

Return the resulting interval graph of an allen computation

evalAllen :: Allen a -> a

Return the result of an allen computation

runAllen :: Allen a -> (a, IntervalGraph)

Return the result of an allen computation and the resulting interval graph